+++ lib/src/scope.dart
@@ -46,9 +46,9 @@ class Scope {
   /// Calls an isolate-scoped RPC named [method] with [params].
   ///
   /// This always adds the `isolateId` parameter to the RPC.
-  Future<Map> sendRequest(String method, [Map<String, Object> params]) async {
+  Future<Object> sendRequest(String method, [Map<String, Object> params]) async {
     var allParams = {"isolateId": isolateId}..addAll(params ?? {});
-    return (await peer.sendRequest(method, allParams)) as Map;
+    return await peer.sendRequest(method, allParams);
   }
 
   /// Evaluates [expression] in the context of the object identified by [id].